move ecdsa.ts to /src/common
