+++ src/Wallet.ts
@@ -2,7 +2,7 @@ import {fromSeed} from 'bip32'
 import {mnemonicToSeedSync} from 'bip39'
 import {decode, encodeForSigning} from 'ripple-binary-codec'
 import {deriveKeypair, generateSeed, verify} from 'ripple-keypairs'
-import ECDSA from './common/types/enums/ecdsa'
+import ECDSA from './common/ecdsa'
 import {SignedTransaction} from './common/types/objects'
 import {signOffline} from './transaction/sign'
 import {SignOptions} from './transaction/types'
